Added the ability to use the builtin function for checking email addresses.  Since this only works with PHP 5.2 and greater, I check to see if the function exists first, otherwise fail-over to the old way.
Also, I went through your regular expression and fixed many issues with it.  [a-z0-9_] with the i modifier is the same as \w.  \w is shorter and easier to read.  You also had the x modifier without any space or comments, so there's not much need for that.  Also, there's no need to escape a + within a character class.  Also, a - at either end of a character class does not denote a range and therefore does not need to be escaped.

The escaping changes are trivial, but the readability of \w over the other way makes it worthwhile.