Fixes language file names having to be unique on a project basis.
Without this you would be unable to have two different languages with the same _lang files in them. For example:

If you had the following folder structure:
  application/language:
    en/general_lang.php
    de/general_lang.php

If you were to load the en general_lang file as default and then try to load the German one on a per-user basis the file wouldn't get loaded because the load function would only check the filename and ignore its 'idiom'. This patch fixes that issue.