Fixed issue with sqlsrv and affected_rows()
Code now does not use a scrollable cursor for UPDATE and INSERT statements. Previously would get "sqlsrv_rows_affected(): supplied resource is not a valid ss_sqlsrv_stmt resource" and "This function only works with statements that are not scrollable" when trying to use $this->db->affected_row() after running an UPDATE or INSERT statement. This has been tested by myself with a SQL Server 2005 database and the latest sqlsrv drivers.