fix backtrace filtering
The backtrace was filtered to remove CI system files,
but the filter was buggy.  It would also filter out
application files which happened to contain the string "system"...
or ALL files, if the application directory is under /system/
(Perhaps the latter comes as a surprise, but it's explicitly
 mentioned in index.php and <http://codeigniter.com/wiki/mod_rewrite>).

Instead, we should test whether the file is underneath BASEPATH
(using realpath() to make sure we have the same sort of slashes).