Try to explain $config['uri_protocol'] better
The main purpose here is to document that REQUEST_URI is a hack.
To be specific, it relies on SCRIPT_NAME.  When using Apache
mod_rewrite as suggested in the CI manual, this only "happens" to work:
it assumes the rewrite only removes one path segment.

Also imply the same about AUTO (which defaults to REQUEST_URI).

Since the non-hacky protocol is PATH_INFO, also outline why PATH_INFO is
not the default - it's because we support versions of PHP which are not
CGI-compliant - and make explicit that ORIG_PATH_INFO is a workaround for
these older versions.  <https://bugs.php.net/bug.php?id=31892>