REQUEST_URI and QUERY_STRING are url-encoded - so decode them
The CGI specification (and testing with PHP) says that QUERY_STRING
is url-encoded.  (PATH_INFO is decoded though).  REQUEST_URI must be
url-encoded otherwise it wouldn't work, and this is confirmed by
testing (Apache).

"the origin server MUST decode the Request-URI in order to
 properly interpret the request" - RFC 2616

"The QUERY_STRING variable contains a URL-encoded search or parameter
 string" - CGI/1.1