URI_test: make sure we clear $_GET before trying to test it
At this point in the test, we're assuming that $_GET is empty.
But we've been calling uri->_fetch_uri_string(), which
(for better or worse) sets $_GET under certain circumstances.

This change will make sure the test will pass when
the implementation of _fetch_uri_string() changes
(see next commit).