Remove obsolete use of getenv()
There's no need to use getenv() for QUERY_STRING (or CONTENT_TYPE).
The only reason to use getenv() was to bypass a bug in PHP's handling
of PATH_INFO.

PHP doesn't mangle the other two.  If they're accessible through
getenv(), they will also be present in $_SERVER.  Every SAPI that
defines getenv() also copies the entire environment into $_SERVER.
(See e.g. php_import_environment_variables() and the CGI SAPI).

And the PATH_INFO bug is fixed in PHP 5.2.4 - which is the minimum
required for this version of CodeIgniter.
<https://bugs.php.net/bug.php?id=31892>