Fix fallback to REQUEST_URI in default configuration
This is my fault.  At some point I misunderstood what isset() does.
Knowing that isset('') = TRUE, I re-reviewed all my changes to URI
parsing and found this issue.

QUERY_STRING could legitimately be "set" to the empty string.
In fact, CGI requires that QUERY_STRING is always set,
even if it would be empty.

So we need to test whether QUERY_STRING is the empty string
before deciding whether to use it (and if it *is* empty,
we will fall back to parsing REQUEST_URI).