Fix request validation for the 'enable_query_strings' case
Under 'enable_query_string' we set the directory, class
and method directly.  Don't smush them into an array
and pass them to _validate_request(), because then
it has to guess whether a directory was used or not.

This commit will also preserve method arguments set in
the 404_override route, if we end up using that.
(Instead of silently ignoring them).