Fix _set_default_controller() not to look in sub-folders
_set_default_controller() sets the class and method directly.
That's ok, because we haven't actually defined what it
would mean to have a default route pointing to a sub-folder.
(Does the URL /directory/ cause us to look in a sub-sub-folder?
Could we get into an infinite loop?)

But then it goes ahead and call _set_request(), which may
try to look in sub-folders.  It would then re-set the class
 - and the method.  Unless the default route doesn't specify
a method, in which case it would set a new class but preserve
the old method, producing an inconsistent request.

Split the part of _set_request() we _do_ want into
_set_request_novalidate(), and call that instead.