added _stringify_attributes() to the helper
this method provides the array-style attribute-specification
interface to the user for attributes of HTML tags.

thanks to github's @ericbarnes, this behaviour already existed in
the _list method. I only abstracted it so it can be used with any
HTML tags; or any future implementations.

currently, this does not break the _list() method it was taken
from and it also works with the heading method. I think this is
enough "testing" for something that already existed in the codebase.