Prevent old multisessions from overwriting current session
The initial behavior of this patch was to have expired multisession
generate a new cookie. The thinking was that once an old session came
in that we would want to ignore it and therefore generate a new cookie.
However, if a request had a significant delay, or if many requests were
coming to the server simultaneously, then from time to time an old
session would hit the server alongside the current session, which
caused a new cookie to be sent back to the client. This resulted in the
session being dropped.

To get around this, expired multi-sessions are no longer longer
forcibly deleted when they are encountered, and garbage collection now
deletes them only when they have passed the session expiration. We
cannot count on networks to behave nicely all the time and it is too
easy to cause dropped sessions by assuming that we won't see a
particular multisession again.

We still need to make sure that we don't allow expired multisessions to
be treated as the current session. To deal with this I've added a
function named "multisession_expired" which returns true if we are
dealing with an expired multisession. This allows us to test a cookie
to see if it is still valid. Note that once a multisession has expired,
we are not longer allowed to change any of the data associated with it.