print_debugger() now accepts three optional parameters to decide what the function returns (Issue #1742)
The function now accepts three boolean parameters for: header, subject
and body (including attachment data). The user can decide what
combination of the above the function should return. The parameters all
have a default value of TRUE, so it is fully backwards compatible with
the previous version of the function.