Remove trailing delimiters from csv_from_result
When using the csv_from_result function, the returned string includes an
extra delimiter at the end of every line, usually a comma unless another
delimiter is specified.  A simple addition of a couple of lines to
remove the extra delimiter from the column names and the data rows is
included. (Lines 241 and 251)