Base URL Functions that Force HTTPS or HTTP
These extend the base URL function by manipulating the URL to have HTTPS
or HTTP.  I needed this since some of my pages needed to be secure and
others didn't.  I used the htaccess file to force the correct connection
and these functions to force the correct links to the secure and
unsecure pages.