Update DB_query_builder.php
Modified join function. If importing the code to an existing CI project, the join will escape the condition is not specified expicitly. This change will now check if the right-handside of the expression is either inside single/double quotes or just numeric and if so it will not escape the RHS with a backtick.