#92 `directory_map()` alternative for `get_dir_file_info()`
- Added optional parameter to `directory_map()` to call
`get_file_info()` on each file and add the result to the function's
output.
- Documented limitation of `get_dir_file_info()` when multiple files in
the directory hierarchy share the same name, and recommended use of
`directory_map()` as a possible solution.

This is not necessarily a fix for #92, but a work-around to maintain
backwards compatibility (in both functions) and make it relatively easy
to gain access to this information under these conditions.