Socket connection for the mysqli driver
If you try to do a socket connection with mysqli you get an error,
because the parameter for that is set in the core to null.

To get it work, i have added a new parameter to the databse config and
did a small change in mysqli_driver.php to allow a socket connection.