Fixed bug - using field_data() on Oracle databases
When you're using oracle databases and want to retrieve column information through the function field_data($table) you get the following notice: 

- Notice: Undefined property: stdClass::$COLUMN_DEFAULT in system/database/drivers/oci8/oci8_driver.php on line 576;

This happens because the oci8 driver tries to access a property that does not exist on query used to get field information. Checking the code we see a small validation to set default value, but the variable $default is not used. So we fix this bug by simply changing:

$retval[$i]->default = $query[$i]->COLUMN_DEFAULT;
to 
$retval[$i]->default = $default;

Bug fixed. No more notices and the properly value is set.