Update postgre_driver.php
Updating function "_execute()" - Currently this function returns a "resource" when the query is performed successfully. When failing it throw a PHP exception (which doesn't happen when running on MySQL). Instead of using "pg_query()" now it uses the function "pg_send_query()", thus saving the result in a variable that will be accessible through the error capture function. If the error code is 0, it means there were no errors and it'd the outcome, otherwise it'd return 'false' (following the same behavior when running on MySQL).

Updating function "error()" - From now saving the result resource it's possible to retrieve what were the query's error Code and Message. Before the error Code was not being taken and the flow diverted due to the exception even the Message was not available.