Respect $config['cur_page'] variable for pagination
After upgrading to CI3 I noticed that developers are able to determine the current page counter for pagination based on

* Explicit query string parameters
* URI segment configuration

In earlier versions a developer could still set the current page counter in the pagination lib directly which is useful if you want to use pagination with HTTP POST instead of GET. This could be done by passing

    $config['cur_page']  = '10';

to the pagination function for link generation. Currently this code has changed and will always try to check whether the uri segment is a valid number or not, even if the cur_page variable was passed in the associative array, and fallback to zero if it fails to validate that result.

This can be easily resolved by checking whether the counter was already set with a valid number from the $config array before trying to resolve it from the uri segment.

This fix give a developer more flexibility and stop CI from overwriting the externally set value with an incorrect one.

Signed-off-by: jekkos <jeroen.peelaerts@gmail.com>