allow multiple validation callbacks for uploads
The current implementation of _execute discards all but the first
validation rule for fields with empty post data. When a field
corresponds to a file-upload field, this means it is not possible to,
for example, handle the file upload in the first callback and perform
validation of the uploaded files in subsequent callbacks.

This PR provides support for cascading callbacks for fields with empty
post data.