Not returning array when there's no DB error
In Oracle Drivers I've noticed that when we call `$this->db->error()` and there's no error, then we get a `false` operator, which allows us to compare in our applications.

When I tested the same result with PostgreSQL I came across a possible bug. Even when there's no error in the query (or even when there's no query executed at all) the function `$this->db->error()` always return a blank array with `code` and `message` keys.

I rewrote the function to check if there's actually an error and return `false` or the array depending on its eval.