Update pdo_driver.php
This  update adds the ability to specify a pdo dsn that is defined in php.ini
For example in the  php.ini file you can say pdo.dsn.<dsn name> = "<connection string>"
And in your application you just specify the dsn name instead of the connection string.  The patch also verifies that the dsn is defined in the photo.ini and if it isn't it defaults back to regular behavior.