Fixed show_error() for check $exit_status
At the beginning of the function it checks whether a parameter $status_code is less than 100. Then the parameter $exit_status is formed by summing the $status_code + 9. In this case, the parameter $exit_status can not be more than 108, but somehow it is checked whether the value of the parameter $exit_status over 125. Such a situation can not occur under any circumstances.