Update DB_driver.php
Add the parameter "refresh_cache" to allow users to refresh the data cache used for listing the table and field data of the database.
This is particularly useful when using the functions like "field_exists" and "table_exists" on migration files.
This is aimed to solve a problem which occurs when using dbforge to make a change to a table or a field and verifying the integrity of the database afterwards due to the fact that these functions are returning information of the database prior to changes.

The following example will show the problem.

Example:

$this->db->field_exists('photo', 'user'); // this will return TRUE
$this->db->field_exists('new_photo', 'user'); // this will return FALSE

$this->dbforge->modify_column('user', array(
	'photo' => array(
		'name' => 'new_photo',
		'type' => 'VARCHAR',
		'constraint' => '256',
		'null' => TRUE
	)
));

$this->db->field_exists('photo', 'user'); // this will return TRUE
$this->db->field_exists('new_photo', 'user'); // this will return FALSE

This happens because the function "field_exists" will get the information previously stored in the variable "data_cache" instead of getting the correct information. This behaviour is correct when using the database normally, but when making changes to the database through migrations is not correct, because you want to know the current state of the database and not the state previously loaded. By using the added parameter it is possible to maintain the normal operation of these functions and at the same time give users the ability to use them safely.