Add single quote to converted character
In many non-english languages, single quote ( ' ) - or apostrophe - is a word separator. It has te be converted into separator (dash or underscore) when the function url_title is called. I updated the regex condition to convert single quotes into separator.

For those that would say : "single quotes must be converted previously", I'll answer that for foreign languages, we already have to convert accented characters. To avoid having too many function calls just for creating a slug, it would be nice to convert the single quote in the url_title function.