Remove NULL check from protected _wh_in function
The NULL check that has been removed was quietly skipping the where_in statement if a NULL was passed in the $value parameter. By not checking for NULL, this allows an inadvertent NULL to get pushed into the query built by the query builder and avoid any accidental updates or selects of the entire table.