Speed up of $this->db->count_all_results()
Count_all_results is currently unnecessarily executing the full prepared select in the query, which might include subqueries that will be executed for each row. 
If the select with no limit is returning a lot of rows it gets slow, as there will be a big amount of queries performed. (As far as i know the working of subqueries, with a select with 20 subqueries, and a no limited return of 100.000 rows, it is 2.000.001 queries plus the count...)

In theory, by setting the select as i propose, you will avoid the unnecessary overwork, as each 'would be row' will simply contain a '1', and there will be only one query executed.

As far as i know, 'distinct' (which to my understanding, is the same as making a 'group by' for each field in the query) queries might return a different number of rows so this quick optimization of the query is not usable with it. In my opinion, it could be posible by converting de 'distinct' to 'group by' for each field selected, feel free to explore that.

Please, check if i missed something.